Creates and outputs a service account key
TL;DR
-----

Fixes an error where the created service account had no keys

Details
-------

Allows us to use the created service account as intended by
creating a service account key for it and making the key available
as part of the Terraform output. Also updated the README to add
a role to the account that runs these templates.